% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesiano.R
\name{cartesiano}
\alias{cartesiano}
\title{Plota ponto (x;y)}
\usage{
cartesiano(x, y)
}
\arguments{
\item{x}{coordenada x}

\item{y}{coordenada y}
}
\value{
ponto (x, y) no pano cartesiano
}
\description{
Plota ponto (x;y)
}
\examples{

cartesiano (3, 2)

}
